+++ dlrover/python/master/node/local_job_manager.py
@@ -10,7 +10,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Dict
 
 from dlrover.python.common.constants import NodeStatus, NodeType
 from dlrover.python.common.grpc import ParallelConfig
@@ -20,10 +19,6 @@ from dlrover.python.master.monitor.error_monitor import (
     SimpleErrorMonitor,
 )
 from dlrover.python.master.node.job_manager import JobManager
-from dlrover.python.master.node.training_node import (
-    SyncNodeTrainingPorts,
-    TrainingNodeConfigure,
-)
 from dlrover.python.scheduler.job import JobArgs
 
 
@@ -156,11 +151,6 @@ class LocalJobManager(JobManager):
     def update_node_paral_config(self, node_type, node_id, paral_config):
         node = self._job_nodes[node_type][node_id]
         node.update_paral_config(paral_config)
-
-    def sync_node_training_port(self, node_id, port) -> SyncNodeTrainingPorts:
-        return self._training_node_configure.sync_node_training_port(
-            node_id, port
-        )
 
 
 def create_job_manager(args: JobArgs, speed_monitor) -> LocalJobManager:
